requirements.txt
bitstring==3.1.9
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
fonttools==4.38.0
gunicorn==20.1.0
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lifxlan==1.2.7
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.24.1
packaging==22.0
Pillow==9.3.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.9.3
six==1.16.0
typer==0.7.0
Werkzeug==2.2.2
